+++ src/main/java/com/my/stock/stockmanager/dto/stock/request/StockSaveRequest.java
@@ -3,11 +3,13 @@ package com.my.stock.stockmanager.dto.stock.request;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class StockSaveRequest {
 	private Long bankId;
 	private String symbol;
-	private Double quantity;
-	private Double price;
+	private BigDecimal quantity;
+	private BigDecimal price;
 }
